refactor: remove dead code and apply .NET 8 conventions for data abstractions
