+++ Agendai/Data/Abstractions/DayCell.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.ObjectModel;
+
+
+namespace Agendai.Data.Abstractions;
+
+public class DayCell
+{
+	public int?                         DayNumber { get; set; }
+	public ObservableCollection<object> Items     { get; set; } = new();
+}
