--- Agendai/Data/Models/DayCell.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.ObjectModel;
-
-
-namespace Agendai.Data.Models;
-
-public class DayCell
-{
-	public int?                         DayNumber { get; set; }
-	public ObservableCollection<object> Items     { get; set; } = new();
-}
