+++ Agendai/Services/views/DayViewService.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using Agendai.Data.Abstractions;
 
 
 namespace Agendai.Services.Views;
