+++ Agendai/ViewModels/Agenda/AgendaWindowViewModel.cs
@@ -10,6 +10,7 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using Agendai.Data.Abstractions;
 
 
 namespace Agendai.ViewModels.Agenda;
